Change account balance type

+++ src/main/java/com/grabtutor/grabtutor/entity/Post.java
@@ -19,7 +19,7 @@ public class Post extends BaseEntity {
 
     String imageUrl;
     String description;
-    int reward;
+    double reward;
     @Builder.Default
     boolean isSolved = false;
     @Builder.Default
